+++ lib/pokedex/pokemon_entry.cpp
@@ -70,9 +70,9 @@ namespace pkmn
         entries_created = true;
     }
 
-    pokemon_entry_t::pokemon_entry_t(uint16_t version_id,
-                                     uint16_t species_id,
-                                     uint16_t form_id)
+    pokemon_entry_t::pokemon_entry_t(int version_id,
+                                     int species_id,
+                                     int form_id)
     {
         CONNECT_TO_DB(db);
         if(not entries_created)
@@ -89,7 +89,7 @@ namespace pkmn
             return;
         }
 
-        uint16_t pokemon_id;
+        int pokemon_id;
         std::ostringstream query_stream;
 
         // Use default form if nothing given
@@ -463,11 +463,11 @@ namespace pkmn
         }
     }
 
-    pokemon_entry_t::pokemon_entry_t(const pkmn::pkstring& version_name,
-                                     const pkmn::pkstring& species_name,
-                                     const pkmn::pkstring& form_name)
+    pokemon_entry_t::pokemon_entry_t(const pkmn::pkstring &version_name,
+                                     const pkmn::pkstring &species_name,
+                                     const pkmn::pkstring &form_name)
     {
-        uint16_t form_id;
+        int form_id;
         if(form_name == "Standard" or form_name == species_name or form_name == "")
             form_id = database::get_species_id(species_name);
         else
